requirements.txt
# requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
sqlalchemy==2.0.9
psycopg2-binary==2.9.6
pydantic==1.10.7
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
bcrypt==4.0.1
jinja2==3.1.2
email-validator==2.0.0
alembic==1.10.3